refactor: single type-guards.js file
